Create student report card
Student report card 
                         Anika Asfi,Rownok Jahan Silvy,
                      Nakshi Saha and Al Amin Shaikh
               Computer Science and Engineering Discipline,
                                     Khulna University
                                    Khulna,Bangladesh


Introduction: Student report card provides us with the information about a student’s name, student id with the marks and gpa etc. 

We made this project using C programming language. We used file to store the information of the students.
Reference: http://mini-projects-in-c.blogspot.com/2011/04/student-report-card.html 



Explanation: First, we use structure to store the information in our code. Struct student st. for every case we input the name roll and the marks of 3 subject. For this we just collected 3 subject numbers. We used input to take the desired input from the user if the user wishes to see the data they just simply call displayall() function. Inside input function we called gradecalculation() function to calculate the desired grade after that we appended the structure to the file. When user wants to see the output simply call the displayall() function which will show the total entry with grade and gpa. We used a standard library <stdlib> header . This header gives us a function called system. Using this function we send a string parameter "cls" which tells the programme to clear the screen we used student.txt file to store the data we  created a structure and saved the structure in the student.txt file . when we need to show the data we simply retrieve the data from the txt file. 

How to run: while running the code user will be asked if he wants to see the result menu or entry or to exit using entry he will store the information of students using writestudent function and user will be shown information with grade and details when user goes to result menu



ALL THE CODE AND PIC ARE GIVEN BELOW: 
https://drive.google.com/open?id=1uwJy7l8o-5-L-I5eCFFRTxfCixY9cEIL
Written By:
1.Anika Asfi (Student ID: 170224)
2.Rownok Jahan Silvy (Student ID: 170234)
3.Nakshi Saha(Student ID: 170236)
4.Al Amin Shaikh (Student ID :160229)